car_price_pro/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.2
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.6.0
matplotlib==3.8.4
seaborn==0.13.2
